requirements.txt
urllib3
jupyterlab
langchain[llm]
pinecone-client[grpc]
openai
python-dotenv
unstructured
arxiv
pymupdf
aider-chat
nest_asyncio
scikit-learn
google-api-python-client
langchain_experimental